# perf(api): optimize nvim_cmd