@lewis6991 2022-07-26
Results are sus so appreciate if any can validate them.