@clason 2022-07-26
Yeah, I cannot reproduce this (macOS M1, after burnin):
* before: 28.4 ms ± 1.7 ms
* after: 27.7 ms ± 0.4 ms