@lewis6991 2022-07-26
Ok, this is the most I can squeeze out of it. The main issue is that each argument needs to be contiguous in memory which means we are forced to re-build the raw command.

25-30% perf improvement isn't bad for -6 LOC.