@clason 2022-07-26
OK, after adding the missing boilerplate from your benchmark snippet, I have the following timings
* master: 148.8 ms ±   0.9 ms
* this PR:   105.5 ms ±   0.5 ms