@ii14 2022-07-26
> Anyone know what the equivalent of `kv_printf(cmdline, "%s", str);` is in terms of `kv_concat[_len]/kv_push` is?

Just `kv_concat`? `kvprintf(cmdline, "%s", s)` should be equivalent to `kv_concat(cmdline, s)` but without going through the printf formatting