@ii14 2022-07-26
What was it complaining about? Maybe strings were not null terminated and you have to pass the length explicitly with `kv_concat_len`?