@lewis6991 2022-07-26
I tried all the variations, and the error isn't very helpful. But I think NUL has something to do with it.