@lewis6991 2022-07-26
@ii14 I'm going to stop here. Feel free to look for other improvements.

For reference, here's the command I've been running for the santizer:

```
ASAN_OPTIONS="detect_leaks=0" UBSAN_OPTIONS="print_stacktrace=1" make CMAKE_EXTRA_FLAGS="-DCLANG_ASAN_UBSAN=ON" functionaltest TEST_FILE=test/functional/api/vim_spec.lua TEST_FILTER='nvim_cmd'
```